requirements.txt
blinker==1.6.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
Flask==3.0.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
pycparser==2.21
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==3.0.0
WTForms==3.0.1
